+++ packages/components-js/tests/e2e/specs/text-field-wrapper.e2e.ts
@@ -66,12 +66,7 @@ describe('text-field-wrapper', () => {
       : '';
     const slottedMessage = useSlottedMessage ? `<span slot="message">Message with a ${link}</span>` : '';
 
-    const attrs = [
-      `state="${state}"`,
-      hasLabel && 'label="Some label"',
-      hasUnit && 'unit="km/h"',
-      maxLength && 'show-character-count="true"',
-    ]
+    const attrs = [`state="${state}"`, hasLabel && 'label="Some label"', hasUnit && 'unit="km/h"']
       .filter((x) => x)
       .join(' ');