+++ packages/components/CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [Unreleased]
 
 #### Added
-- `Text Field Wrapper` now has a `showCharacterCount` property which can be used to **show** the character count when a `maxLength` attribute is set on the wrapped `input`.
-- `Textarea Wrapper` now has a `showCharacterCount` property which can be used to **hide** the character count when a `maxLength` attribute is set on the wrapped `textarea`.
+- `Text Field Wrapper` now has a `showCharacterCount` property which can be used to hide the character count when a `maxLength` attribute is set on the wrapped `input`.
+- `Textarea Wrapper` now has a `showCharacterCount` property which can be used to hide the character count when a `maxLength` attribute is set on the wrapped `textarea`.
 
 ### [2.9.3-rc.0-skeletons] - 2022-03-29